requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# MongoDB Integration
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4

# Async Tasks
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Configuration
python-decouple==3.8

# Security
cryptography==41.0.7

# API Documentation
drf-yasg==1.21.7

# CORS
django-cors-headers==4.3.1

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Monitoring
django-prometheus==2.3.1
